Fix isse with can opener lever
