+++ Assets/Prefabs/Room-Prefabs/CatCafe_Pickle.prefab
@@ -23,7 +23,7 @@ GameObject:
   - 23: {fileID: 23000011274097902}
   - 65: {fileID: 65000013110709340}
   - 54: {fileID: 54000011237477784}
-  - 114: {fileID: 114000012481002510}
+  - 114: {fileID: 114000012264012198}
   m_Layer: 0
   m_Name: CatCafe_Pickle
   m_TagString: Untagged
@@ -107,7 +107,7 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.15263322, y: 0.024579128, z: 0.038399402}
   m_Center: {x: 0.0001062192, y: 0, z: 0}
---- !u!114 &114000012481002510
+--- !u!114 &114000012264012198
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
@@ -115,7 +115,7 @@ MonoBehaviour:
   m_GameObject: {fileID: 1000012385342216}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: dbf28140d151aa94bac6bbd46e8453c7, type: 3}
+  m_Script: {fileID: 11500000, guid: 23a2a2aef2e440647a8dbb33c99c24e8, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   shouldHighlight: 1
