+++ Assets/Scripts/Items/GrabbableObject.cs
@@ -2,9 +2,7 @@
 using System.Collections;
 
 public class GrabbableObject : InteractionBase {
-
-    protected Rigidbody rigidbody;
-    
+ 
     public bool shouldHighlight;
     public Material highlightMaterial;
     protected Material[] savedMaterials;
@@ -25,62 +23,10 @@ public class GrabbableObject : InteractionBase {
 
     protected Transform interationPoint;
 
-    void Start () {
-        rigidbody = GetComponent<Rigidbody>();
-
-        rigidbody.maxAngularVelocity = float.MaxValue;
-
-        AfterStart();
-    }
-
-    public virtual void AfterStart()
-    {
-
-    }
-	
-    void FixedUpdate()
-    {
-        if (rigidbody != null && isGrabbed && controller != null)
-        {
-            float maxDistanceDelta = 10f;
-
-            Quaternion rotationDelta;
-            Vector3 positionDelta;
-
-            float angle;
-            Vector3 axis;
-
-            rotationDelta = controller.transform.rotation * Quaternion.Inverse(interationPoint.rotation);
-            positionDelta = controller.transform.position - interationPoint.position;
-
-            rotationDelta.ToAngleAxis(out angle, out axis);
-
-            angle = (angle > 180 ? angle -= 360 : angle);
-
-            if (angle != 0)
-            {
-                Vector3 angularTarget = angle * axis;
-                rigidbody.angularVelocity = Vector3.MoveTowards(rigidbody.angularVelocity, angularTarget, maxDistanceDelta);
-            }
-
-            Vector3 velocityTarget = positionDelta / Time.fixedDeltaTime;
-            rigidbody.velocity = Vector3.MoveTowards(rigidbody.velocity, velocityTarget, maxDistanceDelta);
-        }
-
-        PostUpdate();
-    }
-
-    public virtual void PostUpdate()
-    {
-
-    }
-
     public void OnGrab(GrabController controller)
     {
         this.controller = controller;
         isGrabbed = true;
-
-        Unfreeze();
 
         if (interationPoint == null)
         {
@@ -186,15 +132,5 @@ public class GrabbableObject : InteractionBase {
                 }
             }
         }
-    }
-
-    public void Freeze()
-    {
-        GetComponent<Rigidbody>().isKinematic = true;
-    }
-
-    public void Unfreeze()
-    {
-        GetComponent<Rigidbody>().isKinematic = false;
     }
 }
