+++ Assets/Scripts/Items/HoldableObject.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 23a2a2aef2e440647a8dbb33c99c24e8
+timeCreated: 1479319700
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
